Fixing an issue with GetSaltFromPass
Upon debugging I found that the GetSaltFromPass function caused Anope to crash if faced with a password that does not match the format specified. In order to prevent a crash the simple error catcher from the function called directly above GetSaltFromPass was applied. If faced with an unconverted or false format password it will now no longer crash but instead print out "Password incorrect." and services will not crash. NOTE: This is not true error handling, this is simply preventing a crash. PS. I did the debugging and the error report, the C++ was done by Adam.