Adding the rest of the checks
GetSaltFromPass is now a bool instead of a void in order for it to return either true or false to make the simple check added work properly. If it does not find a password in the right format it will not return false instead of crashing with an unhandled exception as every string search is now inside a condition that will return false if it fails. The function now returns true.